Move Error renderer to UiBundle
