+++ src/Sylius/Bundle/UiBundle/Resources/config/services/twig/error_renderer.xml
@@ -13,13 +13,16 @@
 
 <container xmlns="http://symfony.com/schema/dic/services" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
     <services>
-        <defaults public="true" />
-
-        <service id="sylius_admin.twig.error_template_finder" class="Sylius\Bundle\AdminBundle\Twig\ErrorTemplateFinder\ErrorTemplateFinder">
-            <argument type="service" id="Sylius\Bundle\CoreBundle\SectionResolver\SectionProviderInterface" />
-            <argument type="service" id="request_stack" />
+        <service
+            id="sylius_ui.twig.error_renderer"
+            class="Sylius\Bundle\UiBundle\Twig\ErrorRenderer\TwigErrorRenderer"
+            decorates="twig.error_renderer.html"
+            decoration-priority="64"
+        >
+            <argument type="service" id="sylius_ui.twig.error_renderer.inner"/>
             <argument type="service" id="twig" />
-            <tag name="sylius.twig.error_template_finder" />
+            <argument type="tagged_iterator" tag="sylius.twig.error_template_finder" />
+            <argument>%kernel.debug%</argument>
         </service>
     </services>
 </container>
